Tons of fixes
