+++ src/WraithXCOD/WraithXCOD/CoDCDNDownloaderV1.cpp
@@ -13,7 +13,7 @@
 #include "WinFileSystem.h"
 
 // We need the XSUB Classes
-#include "VGXSUBCache.h"
+#include "XSUBCacheV2.h"
 #include "XSUBCache.h"
 
 // We need web class
@@ -106,7 +106,7 @@ std::unique_ptr<uint8_t[]> CoDCDNDownloaderV1::ExtractCDNObject(uint64_t cacheID
 
 	if (cdnBuffer != nullptr)
 	{
-		auto decompressed = VGXSUBCache::DecompressPackageObject(
+		auto decompressed = XSUBCacheV2::DecompressPackageObject(
 			entry.Hash,
 			cdnBuffer.get(),
 			bufferSize,
@@ -140,7 +140,7 @@ std::unique_ptr<uint8_t[]> CoDCDNDownloaderV1::ExtractCDNObject(uint64_t cacheID
 	// Add to the cache
 	Cache.Add(cacheID, result->DataBuffer, result->BufferSize);
 
-	return VGXSUBCache::DecompressPackageObject(
+	return XSUBCacheV2::DecompressPackageObject(
 		entry.Hash,
 		result->DataBuffer,
 		result->BufferSize,
