+++ src/WraithXCOD/WraithXCOD/WraithXCOD.rc
@@ -337,7 +337,7 @@ IDB_SETTINGSHOVER       PNG                     "..\\Resources\\SettingHover.png
 
 // Icon with lowest ID value placed first to ensure application icon
 // remains consistent on all systems.
-IDI_MAINICON            ICON                    "..\\Resources\\GreyhoundEnhanced.ico"
+IDI_MAINICON            ICON                    "..\\Resources\\Greyhound2023.ico"
 
 IDI_CHECKMARK           ICON                    "..\\..\\WraithX\\Resources\\CheckMarkIcon.ico"
 
@@ -348,8 +348,8 @@ IDI_CHECKMARK           ICON                    "..\\..\\WraithX\\Resources\\Che
 //
 
 VS_VERSION_INFO VERSIONINFO
- FILEVERSION 1,30,0,0
- PRODUCTVERSION 1,30,0,0
+ FILEVERSION 1,40,0,0
+ PRODUCTVERSION 1,40,0,0
  FILEFLAGSMASK 0x3fL
 #ifdef _DEBUG
  FILEFLAGS 0x1L
@@ -366,12 +366,12 @@ BEGIN
         BEGIN
             VALUE "CompanyName", "DTZxPorter / Scobalula"
             VALUE "FileDescription", "Greyhound"
-            VALUE "FileVersion", "1.30.0.0"
+            VALUE "FileVersion", "1.40.0.0"
             VALUE "InternalName", "Greyhound"
             VALUE "LegalCopyright", "Copyright (C) 2022 DTZxPorter / Scobalula"
             VALUE "OriginalFilename", "Greyhound.exe"
             VALUE "ProductName", "Greyhound"
-            VALUE "ProductVersion", "1.30.0.0"
+            VALUE "ProductVersion", "1.40.0.0"
         END
     END
     BLOCK "VarFileInfo"
