+++ src/WraithXCOD/WraithXCOD/XSUBCacheV2.h
@@ -2,7 +2,7 @@
 // We need the package cache
 #include "CoDAssets.h"
 #include "CoDPackageCache.h"
-#include "CoDFileSystem.h"
+#include "Casc.h"
 #include <shared_mutex>
 
 // A class that handles reading, caching and extracting CASC Resources
@@ -19,10 +19,13 @@ public:
     virtual bool LoadPackage(const std::string& FilePath);
     // Implement the extract function
     virtual std::unique_ptr<uint8_t[]> ExtractPackageObject(uint64_t CacheID, int32_t Size, uint32_t& ResultSize);
+
+    // Decompresses a compressed package object.
+    static std::unique_ptr<uint8_t[]> DecompressPackageObject(uint64_t cacheID, uint8_t* buffer, size_t bufferSize, size_t decompressedSize, size_t& resultSize);
 };
 
 #pragma pack(push, 1)
-struct XSUBBlockV2
+struct VGXSUBBlock
 {
     uint8_t Compression;
     uint32_t CompressedSize;
@@ -34,7 +37,7 @@ struct XSUBBlockV2
 #pragma pack(pop)
 
 #pragma pack(push, 1)
-struct XSUBHeaderV2
+struct VGXSUBHeader
 {
     uint32_t Magic;
     uint16_t Unknown1;
@@ -43,29 +46,28 @@ struct XSUBHeaderV2
     uint64_t Type;
     uint64_t Size;
     uint8_t UnknownHashes[1896];
-    uint64_t FileCount;
-    uint64_t DataOffset;
-    uint64_t DataSize;
-    uint64_t HashCount;
-    uint64_t HashOffset;
-    uint64_t HashSize;
-    uint64_t Unknown3;
-    uint64_t UnknownOffset;
-    uint64_t Unknown4;
-    uint64_t IndexCount;
-    uint64_t IndexOffset;
-    uint64_t IndexSize;
+    int64_t FileCount;
+    int64_t DataOffset;
+    int64_t DataSize;
+    int64_t HashCount;
+    int64_t HashOffset;
+    int64_t HashSize;
+    int64_t Unknown3;
+    int64_t UnknownOffset;
+    int64_t Unknown4;
+    int64_t IndexCount;
+    int64_t IndexOffset;
+    int64_t IndexSize;
 };
 #pragma pack(pop)
 
 #pragma pack(push, 1)
-struct XSUBHashEntryV2
+struct VGXSUBHashEntry
 {
     uint64_t Key;
     uint64_t PackedInfo;
-    uint32_t PackedInfoEx;
 };
 #pragma pack(pop)
 
 // Verify
-static_assert(sizeof(XSUBBlockV2) == 0x15, "Invalid Vanguard Block Struct Size (Expected 0x15)");
+static_assert(sizeof(VGXSUBBlock) == 0x15, "Invalid Vanguard Block Struct Size (Expected 0x15)");
