+++ src/WraithXCOD/WraithXCOD/XSUBCacheV3.h
@@ -2,21 +2,16 @@
 // We need the package cache
 #include "CoDAssets.h"
 #include "CoDPackageCache.h"
-#include "Casc.h"
+#include "CoDFileSystem.h"
 #include <shared_mutex>
 
 // A class that handles reading, caching and extracting CASC Resources
-class VGXSUBCache : public CoDPackageCache
+class XSUBCacheV3 : public CoDPackageCache
 {
-private:
-    // Container
-    Casc::Container Container;
-    // Cache Mutex
-    std::shared_mutex ReadMutex;
 public:
     // Constructors
-    VGXSUBCache();
-    virtual ~VGXSUBCache();
+    XSUBCacheV3();
+    virtual ~XSUBCacheV3();
 
     // Implement the load function
     virtual void LoadPackageCache(const std::string& BasePath);
@@ -24,13 +19,10 @@ public:
     virtual bool LoadPackage(const std::string& FilePath);
     // Implement the extract function
     virtual std::unique_ptr<uint8_t[]> ExtractPackageObject(uint64_t CacheID, int32_t Size, uint32_t& ResultSize);
-
-    // Decompresses a compressed package object.
-    static std::unique_ptr<uint8_t[]> DecompressPackageObject(uint64_t cacheID, uint8_t* buffer, size_t bufferSize, size_t decompressedSize, size_t& resultSize);
 };
 
 #pragma pack(push, 1)
-struct VGXSUBBlock
+struct XSUBBlockV2
 {
     uint8_t Compression;
     uint32_t CompressedSize;
@@ -42,7 +34,7 @@ struct VGXSUBBlock
 #pragma pack(pop)
 
 #pragma pack(push, 1)
-struct VGXSUBHeader
+struct XSUBHeaderV2
 {
     uint32_t Magic;
     uint16_t Unknown1;
@@ -51,28 +43,29 @@ struct VGXSUBHeader
     uint64_t Type;
     uint64_t Size;
     uint8_t UnknownHashes[1896];
-    int64_t FileCount;
-    int64_t DataOffset;
-    int64_t DataSize;
-    int64_t HashCount;
-    int64_t HashOffset;
-    int64_t HashSize;
-    int64_t Unknown3;
-    int64_t UnknownOffset;
-    int64_t Unknown4;
-    int64_t IndexCount;
-    int64_t IndexOffset;
-    int64_t IndexSize;
+    uint64_t FileCount;
+    uint64_t DataOffset;
+    uint64_t DataSize;
+    uint64_t HashCount;
+    uint64_t HashOffset;
+    uint64_t HashSize;
+    uint64_t Unknown3;
+    uint64_t UnknownOffset;
+    uint64_t Unknown4;
+    uint64_t IndexCount;
+    uint64_t IndexOffset;
+    uint64_t IndexSize;
 };
 #pragma pack(pop)
 
 #pragma pack(push, 1)
-struct VGXSUBHashEntry
+struct XSUBHashEntryV2
 {
     uint64_t Key;
     uint64_t PackedInfo;
+    uint32_t PackedInfoEx;
 };
 #pragma pack(pop)
 
 // Verify
-static_assert(sizeof(VGXSUBBlock) == 0x15, "Invalid Vanguard Block Struct Size (Expected 0x15)");
+static_assert(sizeof(XSUBBlockV2) == 0x15, "Invalid Vanguard Block Struct Size (Expected 0x15)");
